Refactor TypedExpr normalization
